[Table/Create/240325] -Modify datatypes

+++ database/migrations/2024_03_23_072225_create_order_lines_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('order_id');
             $table->unsignedBigInteger('qty');
-            $table->unsignedBigInteger('price');
+            $table->double('price');
             $table->timestamps();
 
             // Foreign Keys
