+++ database/migrations/2024_03_23_072912_create_shipping_method_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('shipping_method', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique()->nullable();
-            $table->unsignedBigInteger('price');
+            $table->double('price');
             $table->timestamps();
         });
     }
